Move Builder to the listener namespace.
